requirements.txt

Flask
Flask-RESTful
Flask-JWT-Extended
python-dotenv
boto3
psycopg2-binary

Flask
Flask-RESTful
Flask-JWT-Extended
python-dotenv
boto3
psycopg2-binary
